==> benchmark pipeline/01-BenchmarkUsability.Rmd <==
---
title: "01-BenchmarkUsability.Rmd"
output: html_document
date: "2024-12-30"
editor_options: 
  chunk_output_type: console
license: >
  Copyright 2024 Chuhan Wang, Adam S. Chan, Xiaohang Fu, Shila Ghazanfar, Jinman Kim, Ellis Patrick, Jean YH Yang. Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0. Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This script displays the code used to generate usability plot and generate usability score rankings for each method

### Load functions required

```{r}
suppressPackageStartupMessages({
  library(rlang)
  library(dplyr)
  library(tidyr)
  library(readxl)
  library(ggplot2)
  library(patchwork)
  library(ggthemr)
})

select <- dplyr::select
slice <- dplyr::slice
# Set ggplot theme
ggthemr_reset()
ggthemr('pale')
mycolors <- c("#999999", "#0072B2", "#E69F00", "#F0E442", "#D9B3FF", "#009E73",  
              "#D55E00", "#5D8AA8", "#CC79A7", "#56B4E9",
              "#F3B3A6", "#A5AB81", "#B2182B", "#4393C3", "#CDBE6B", 
              "#80CDC1", "#F4A582", "#BABABA", "#CCEBC5", "#DECBE4",
              "#FDDFDF", "#B3DE69", "#FDBF6F", "#CCECE6", "#FB8072")

mypalette <- define_palette(
  swatch = mycolors,
  gradient = c(lower = mycolors[1L], upper = mycolors[3L])
)

ggthemr(mypalette) # for some reason it uses the first colour as the colour of the grid lines
theme_update(panel.grid.major = element_line(linetype="dotted"))
```

### Load Data

The following code imports the usability data generated by surveying each of the analysed methods. 

```{r}
method_vars <- c("GeneCodeR", "ST-Net", "DeepPT",
                 "Hist2ST","HisToGene","DeepSpaCE","EGNv1", "EGNv2", "TCGN", "THItoGene", "iStar")

usability_df <- read_excel('data/raw/methods_usability_scoring.xlsx') %>%
  fill(name, aspect_id, category, aspect) %>%
  filter(!(aspect_id %in% c("license","seed_setting"))) %>%
  # Remove last row
  slice(-nrow(.)) %>%
  slice(nrow(.):1) %>%
  mutate_at(vars(all_of(method_vars)),
            as.numeric)  %>%
  mutate(y_end = cumsum(item_weight), 
         y_start = c(0, y_end[-nrow(.)]))
```

Code for generating Figure 5 of the manuscript

```{r fig.height=9, fig.width=7}
# Create a df 
usability_by_names <- usability_df %>%
  group_by(name) %>%
  summarise(y_mid = (min(y_start)+ max(y_end))/2) 

use_score_df <- data.frame(
  method = method_vars,
  overall_score = lapply(
    method_vars,
    function(col) {
      sum(usability_df$item_weight * usability_df[,col])
    }
  ) %>% 
    unlist()
) %>%
  arrange(-overall_score)

p_method_score <- use_score_df %>%
  mutate(method = factor(method, levels=use_score_df$method)) %>%
  ggplot() +
  aes(x=method, y=overall_score, fill=overall_score) +
  geom_bar(stat='identity') +
  geom_text(aes(label=round(overall_score,2)),
            vjust=2,
            size=2.3, col="grey50") +
  theme_void() +
  scale_fill_gradient2(high="#75a98f", low="#BCD6C9") +
  labs(fill="Usability Score")

p_score_hm <- usability_df %>%
  select(item, item_weight, category,
         y_end, y_start, all_of(method_vars)) %>%
  pivot_longer(cols=!c("item","item_weight", "category",
                       "y_end", "y_start"),
               names_to="method",
               values_to="score") %>%
  mutate(method = factor(method, levels=use_score_df$method),
         category = factor(category, levels=rev(unique(.$category))),
         score = factor(score, levels=c(0,.25,.5,.75,1))) %>%
  ggplot() +
  aes(ymin=y_start, ymax=y_end, x=method, fill=score) +
  geom_blank()+
  geom_rect(aes(xmin=as.numeric(method)-0.5, xmax=as.numeric(method)+0.5),
            col="white") +
  scale_fill_manual(values = c("grey80", "grey65", "grey50", "grey35", "grey20"), 
                    breaks = c(0, 0.25, 0.5, 0.75, 1)) +
  scale_y_continuous(expand = c(0, 0),
                     breaks=usability_by_names$y_mid,
                     labels=usability_by_names$name) +
  theme(strip.placement = "outside",
        strip.text = element_text(size=7),
        strip.background = element_rect(fill="grey85"), 
        axis.text.x = element_text(angle=45,hjust=1,vjust=1),
        panel.background = element_blank(),
        axis.line.x = element_blank()) +
  facet_grid(rows=vars(category), 
             scales="free_y", switch = "y") +
  labs(fill="Score", x="Method")

p_use_score <- usability_df %>%
  mutate(category = factor(category, levels=rev(unique(.$category)))) %>%
  mutate(use_score = rowMeans(usability_df[,method_vars], na.rm=TRUE)) %>%
  ggplot() +
  aes(ymin=y_start, ymax=y_end, fill=use_score) +
  geom_rect(aes(xmin=0, xmax=1),
            col="gray") +
  scale_fill_gradient2(high="#5ccdcd", low="#e29d9d", mid="#dedad2",
                       midpoint=0.5) +
  scale_y_continuous(expand = c(0, 0)) +
  scale_x_continuous(expand = c(0, 0)) +
  facet_grid(rows=vars(category), 
             scales="free_y", switch = "y") +
  theme_void() +
  theme(axis.ticks = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        strip.background = element_blank(),
        strip.text = element_blank()) +
  labs(fill="Average Score")

plt <- p_method_score + plot_spacer() + p_score_hm + p_use_score +
  plot_layout(widths = c(9, 1), heights=c(1,5), guides = 'collect')

plt
```

### Save data

This saves the ranked data required for overall benchmark figure

```{r}
use_cat_score_df <- usability_df %>% 
  pivot_longer(cols=c("GeneCodeR", "ST-Net", "DeepPT", "Hist2ST", "HisToGene", "DeepSpaCE", "EGNv1", "EGNv2", "TCGN", "THItoGene", "iStar"), 
  names_to="model_id", values_to="score") %>% 
  group_by(category, model_id) %>% 
  summarise(mean_score = mean(score)) %>% 
  pivot_wider(names_from="category", values_from="mean_score")

save(use_score_df,
     use_cat_score_df,
     file='data/processed/methods_usability_dat.RData')
```
